\name{datamed}
\alias{datamed}
\docType{data}
\title{
Example data
}
\description{
Example data; simulated
}
\usage{data("datamed")}
\format{
  A data frame with 5000 observations on the following 7 variables.
  \describe{
    \item{\code{outcome}}{the binary outcome,a numeric vector}
    \item{\code{exposure}}{the binary exposure, a numeric vector}
    \item{\code{med1}}{mediator 1, a numeric vector}
    \item{\code{med2}}{mediator 2, a numeric vector}
    \item{\code{med3}}{mediator 3,a numeric vector}
    \item{\code{con1}}{confounder 1, a numeric vector}
    \item{\code{con2}}{confounder 2, a numeric vector}
  }
}
